refactor: :truck: Rename home bloc files
